Middleware added, it works, also i refactored the code of routes.js a little bit.
