+++ app/controllers/usersController.js
@@ -1,20 +1,9 @@
 const { registerUser } = require('../services/users');
-const { validate } = require('../validators/users');
 const { encryptPassword } = require('../helpers/encryption');
 const logger = require('../logger/index');
 
 const addUser = (req, res) => {
   const user = req.body;
-  const validationErrors = validate(user).errors;
-
-  if (validationErrors.length) {
-    logger.error(
-      `The input user data: ${user.firstName} ${user.lastName} ${user.email} is not valid: ${validationErrors}`
-    );
-    return res.status(500).send({
-      error: `The input user data: ${user.firstName} ${user.lastName} ${user.email} is not valid: ${validationErrors}`
-    });
-  }
 
   user.password = encryptPassword(user.password);
