+++ app/routes.js
@@ -1,10 +1,15 @@
 /* eslint-disable new-cap */
 
 const express = require('express');
-const albumsController = require('./controllers/albumsController');
-const photosController = require('./controllers/photosController');
-const usersController = require('./controllers/usersController');
-const validate = require('./middlewares/validations');
+const {
+  showAllAlbums,
+  showAlbumById,
+  showPhotosFromAlbum,
+  showPhotoFromAlbumByIds
+} = require('./controllers/albumsController');
+const { showAllPhotos, showPhotoById } = require('./controllers/photosController');
+const { addUser } = require('./controllers/usersController');
+const { validation } = require('./middlewares/validations');
 const { healthCheck } = require('./controllers/healthCheck');
 
 const init = app => {
@@ -13,18 +18,18 @@ const init = app => {
 
 const photosRouter = express.Router();
 
-photosRouter.get('/', photosController.showAllPhotos);
-photosRouter.get('/:id', photosController.showPhotoById);
+photosRouter.get('/', showAllPhotos);
+photosRouter.get('/:id', showPhotoById);
 
 const albumsRouter = express.Router();
 
-albumsRouter.get('/', albumsController.showAllAlbums);
-albumsRouter.get('/:id', albumsController.showAlbumById);
-albumsRouter.get('/:id/photos/', albumsController.showPhotosFromAlbum);
-albumsRouter.get('/:idAlbum/photos/:idPhoto', albumsController.showPhotoFromAlbumByIds);
+albumsRouter.get('/', showAllAlbums);
+albumsRouter.get('/:id', showAlbumById);
+albumsRouter.get('/:id/photos/', showPhotosFromAlbum);
+albumsRouter.get('/:idAlbum/photos/:idPhoto', showPhotoFromAlbumByIds);
 
 const usersRouter = express.Router();
 
-usersRouter.post('/', [validate], usersController.addUser);
+usersRouter.post('/', validation, addUser);
 
 module.exports = { albumsRouter, photosRouter, usersRouter, init };
